+++ tests/integ/snowflake/ml/feature_store/feature_store_case_sensitivity_test.py
@@ -95,7 +95,7 @@ class FeatureStoreCaseSensitivityTest(parameterized.TestCase):
         e = Entity(name="e", join_keys=["a"])
         fs.register_entity(e)
         fv = FeatureView(name="fv", entities=[e], feature_df=df, refresh_freq="1 minute")
-        fs.register_feature_view(feature_view=fv, version="v1", block=True)
+        fs.register_feature_view(feature_view=fv, version="v1")
         retrieved_fv = fs.get_feature_view(name="fv", version="v1")
         self.assertEqual(resolve_identifier(database), retrieved_fv.database)
         self.assertEqual(resolve_identifier(schema), retrieved_fv.schema)
@@ -128,7 +128,7 @@ class FeatureStoreCaseSensitivityTest(parameterized.TestCase):
         e = Entity(name="e", join_keys=["a"])
         fs.register_entity(e)
         fv = FeatureView(name="fv", entities=[e], feature_df=df, refresh_freq="1 minute")
-        fs.register_feature_view(feature_view=fv, version="v1", block=True)
+        fs.register_feature_view(feature_view=fv, version="v1")
         retrieved_fv = fs.get_feature_view(name="fv", version="v1")
         self.assertEqual(resolve_identifier(warehouse), retrieved_fv.warehouse)
         self.assertEqual(2, len(fs.read_feature_view(retrieved_fv).to_pandas()))
@@ -254,7 +254,10 @@ class FeatureStoreCaseSensitivityTest(parameterized.TestCase):
             e = Entity(name="MY_COOL_ENTITY", join_keys=[test_name])
             fv_0 = FeatureView(name="MY_FV", entities=[e], feature_df=df, timestamp_col=test_name)
             fs.register_entity(e)
-            fv_1 = fs.register_feature_view(fv_0, "V1", block=True)
+            fv_1 = fs.register_feature_view(
+                fv_0,
+                "V1",
+            )
 
             retrieved_e = fs.get_entity("MY_COOL_ENTITY")
             self.assertEqual(len(retrieved_e.join_keys), 1)
@@ -311,7 +314,7 @@ class FeatureStoreCaseSensitivityTest(parameterized.TestCase):
         original_fv_name, original_version = equi_full_names[0]
         fv_0 = FeatureView(name=original_fv_name, entities=[e], feature_df=df)
         fs.register_entity(e)
-        fs.register_feature_view(fv_0, original_version, block=True)
+        fs.register_feature_view(fv_0, original_version)
 
         # register with identical full name will fail
         for equi_full_name in equi_full_names:
@@ -319,14 +322,14 @@ class FeatureStoreCaseSensitivityTest(parameterized.TestCase):
             version = equi_full_name[1]
             with self.assertRaisesRegex(ValueError, "FeatureView .* already exists"):
                 fv = FeatureView(name=fv_name, entities=[e], feature_df=df)
-                fs.register_feature_view(fv, version, block=True)
+                fs.register_feature_view(fv, version)
 
         # register with different full name is fine
         for diff_full_name in diff_full_names:
             fv_name = diff_full_name[0]
             version = diff_full_name[1]
             fv = FeatureView(name=fv_name, entities=[e], feature_df=df)
-            fv = fs.register_feature_view(fv, version, block=True)
+            fv = fs.register_feature_view(fv, version)
             fs.read_feature_view(fv)
 
         self.assertEqual(len(fs.list_feature_views(as_dataframe=False)), len(diff_full_names) + 1)
@@ -403,28 +406,28 @@ class FeatureStoreCaseSensitivityTest(parameterized.TestCase):
         fv = FeatureView(name="MY_FV", entities=[e], feature_df=df)
 
         # 1: register with lowercase, get it back with lowercase/uppercase
-        fs.register_feature_view(fv, "a1", block=True)
+        fs.register_feature_view(fv, "a1")
         fs.get_feature_view("MY_FV", "A1")
         fs.get_feature_view("MY_FV", "a1")
 
         # 2: register with uppercase, get it back with lowercase/uppercase
-        fs.register_feature_view(fv, "B2", block=True)
+        fs.register_feature_view(fv, "B2")
         fs.get_feature_view("MY_FV", "b2")
         fs.get_feature_view("MY_FV", "B2")
 
         # 3. register with valid characters
-        fs.register_feature_view(fv, "V2_1", block=True)
+        fs.register_feature_view(fv, "V2_1")
         fs.get_feature_view("MY_FV", "v2_1")
-        fs.register_feature_view(fv, "3", block=True)
+        fs.register_feature_view(fv, "3")
         fs.get_feature_view("MY_FV", "3")
 
         # 4: register with invalid characters
         with self.assertRaisesRegex(ValueError, ".* is not a valid feature view version.*"):
-            fs.register_feature_view(fv, "abc$", block=True)
+            fs.register_feature_view(fv, "abc$")
         with self.assertRaisesRegex(ValueError, ".* is not a valid feature view version.*"):
-            fs.register_feature_view(fv, "abc#", block=True)
+            fs.register_feature_view(fv, "abc#")
         with self.assertRaisesRegex(ValueError, ".* is not a valid feature view version.*"):
-            fs.register_feature_view(fv, '"abc"', block=True)
+            fs.register_feature_view(fv, '"abc"')
 
 
 if __name__ == "__main__":