+++ CHANGELOG.md
@@ -1,12 +1,40 @@
 # Release History
 
-## 1.3.1
+## 1.4.0
+
+### Bug Fixes
+
+- Registry: Fix a bug when multiple models are being called from the same query, models other than the first one will
+  have incorrect result. This fix only works for newly logged model.
+- Modeling: When registering a model, only method(s) that is mentioned in `save_model` would be added to model signature
+in SnowML models.
+- Modeling: Fix a bug that when n_jobs is not 1, model cannot execute methods such as
+predict, predict_log_proba, and other batch inference methods. The n_jobs would automatically
+set to 1 because vectorized udf currently doesn't support joblib parallel backend.
+- Modeling: Fix a bug that batch inference methods cannot infer the datatype when the first row of data contains NULL.
+- Modeling: Matches Distributed HPO output column names with the snowflake identifier.
+- Modeling: Relax package versions for all Distributed HPO methods if the installed version
+  is not available in the Snowflake conda channel
+- Modeling: Add sklearn as required dependency for LightGBM package.
+
+### Behavior Changes
+
+- Registry: `apply` method is no longer by default logged when logging a xgboost model. If that is required, it could
+  be specified manually when logging the model by `log_model(..., options={"target_methods": ["apply", ...]})`.
+
+### New Features
+
+- Registry: Add support for `sentence-transformers` model (`sentence_transformers.SentenceTransformer`).
+- Registry: Now version name is no longer required when logging a model. If not provided, a random human readable ID
+  will be generated.
+
+## 1.3.1 (2024-03-21)
 
 ### New Features
 
 - FileSet: `snowflake.ml.fileset.sfcfs.SFFileSystem` can now be used in UDFs and stored procedures.
 
-## 1.3.0
+## 1.3.0 (2024-03-12)
 
 ### Bug Fixes